pyproject.toml
[tool.poetry]
name = "visa-me-ajuda"
version = "0.1.0"
description = ""
authors = ["talesmm14 <talesmelquiades@hotmail.com>"]
readme = "README.md"
packages = [{include = "visa_me_ajuda"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
pyngrok = "^5.1.0"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
